Split span tests
